move SearchBar to 'src/components' folder
