+++ src/components/SearchBar/SearchBar.scss
